Criação de produto criada

+++ src/main/java/math/projeto3/models/ShoppingProductModel.java
@@ -45,11 +45,11 @@ public class ShoppingProductModel {
         this.product = product;
     }
 
-    public Double getQuantity() {
+    public Integer getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Double quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
 }
